+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GBV: Added ordering to I'm Not Sure Yet - Support a Friend results [#5434](https://github.com/rokwire/illinois-app/issues/5434).
 ### Fixed
 - Fixed "Clear All" header bar button text style in Visited Recently panel [#5443](https://github.com/rokwire/illinois-app/issues/5443).
+### Changed
+- Elevate deeplinks in Illinois Assistant Results [#5459](https://github.com/rokwire/illinois-app/issues/5459).
 
 ## [7.3.14] - 2025-10-17
 ### Changed